+++ Tasks/Hatsak/Task3/tests.py
